HABot: fix nullness
After adding the nullness annotations to the automation the type
argument for the base module handler must be a non nullable one.
To apply the non-null annotation to the given type argument we mark the
whole class as non null by default.

Related to: https://github.com/openhab/openhab-core/pull/910

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>