Add default UI bundle (#178)
This add the UI in its current state of development,
the intent is to test the integration in the distro
early. The other apps are listed on the home page by
calling the new REST endpoint introduced in
https://github.com/openhab/openhab-core/pull/1329

The onboarding wizard implementation is not complete
so it isn't possible to select a startup package with
the UI yet.  The wizard steps will be discussed in a
separate issue; it should however be possible to install
addons as needed from the settings screen.

Signed-off-by: Yannick Schaus <github@schaus.net>