Make semantic cards on the home page functional
Use the default list widgets to render items in the card.
(order not considered for now, but visible and visibleTo
configured in the listWidget metadata might work to filter
out items based on an expression/the current user's roles?)

Remove warning messages.

Make cards (and headers) a little bit thinner.

Signed-off-by: Yannick Schaus <github@schaus.net>