[WIP] Cell Widgets
This introduces a new first-level class of widgets in the
standard library: cells.

They are heavily inspired by
https://developer.android.com/preview/features/device-control
and can either perform an action on a click/short tap, like
toggling a switch or perform something stateless like opening
a modal, navigating, running a rule...

It will also expand on a long tap (if the device doesn't support
touch, a button will have to be added). If no short tap/click
action is configured it will default to expanding the cell.

Ways to alter the design of the cell button (when not expanded)
and the expanded controls, eventually with personal widgets
configured in the developer tools, are also on the roadmap.
This will allow the community to develop specialized widgets
for e.g. thermostats, weather forecasts etc.

A series of cells tailored to common scenarios (toggle +
dimmer or color picker) and maybe a few more are envisioned.
There will be a way to define the default cell widget for an
item using the "cellWidget" metadata namespace, in the same
manner as the existing default standalone and list item widgets.

The container, oh-cells, can be added to a block in a layout
page, instead of a regular row. It will automatically arrange
the cells, which have a fixed height and are all of the same
size.

Signed-off-by: Yannick Schaus <github@schaus.net>