Validate parameter immediately & before save
Perform the validation of all configuration parameters
when they are displayed (for instance this will show the
"Please fill out this field" message on all required
parameters immediately when the sheet is displayed).

Validate ID and label inputs on general thing settings.

Validate configuration on thing creation or update, and
profile configuration on link add/edit, and refuse to
save if there are invalid parameters.
Closes #267.

Signed-off-by: Yannick Schaus <github@schaus.net>