Widgets: context variables, add oh-input
This introduces a way to set page-local or widget-local
variables to allow more complex scenarios in pages & custom
widget development.

Variables can be used using several methods:

- the "variable" config parameter of oh-gauge (read-only),
  oh-input (new), oh-knob, oh-slider, oh-stepper, oh-toggle
  will accept a variable name and control it instead of
  sending commands to items if set. The "item" parameter can
  still be set to set the widget to the item's state, when
  the variable has no value.
- the "vars" object available in expressions (for example
  `=vars.var1` will evaluate to the value of the variable `var1`.
- the "variable" action allows to set an fixed or computed
  (using an expression) value to a variable.

Variables set directly by widgets on a page will propagate
to all components to the page, including personal widgets.
However, variables set within a personal widget will be scoped
to that personal widget only - except when there's already an
inherited variable with the same name.

oh-button & oh-link have a special parameter "clearVariable"
which allows to unset a version when clicked, after performing
the action. This is useful when "validating" a variable e.g.
set a command to an item with the variable value then reset it.

Allow oh-image to perform an action when clicked directly.

Signed-off-by: Yannick Schaus <github@schaus.net>