Improve text parameter autocomplete
For config parameters with options and limitToOptions=false,
currently implemented as a text field with autocomplete.

Display & allow filtering on the label, and display all
the available options on open.

Fixes #357.

Don't perform the initial validation pass on config sheet.
Adds a mention in the description for required fields instead.

Closes #353.

Signed-off-by: Yannick Schaus <github@schaus.net>