Channel list improvements
Add `searchbar-ignore` class to link skeletons to prevent
the height of the accordion to be initialized at 0px.
Should fix #401.

Add filters to display only linked or unlinked channels.
Closes #395.

Hide item linking controls for trigger channels.
Closes #390.

Add option to select/unselect all checkboxes when in
multiple links mode. This will take the linked/unlinked
filter into account, and the advanced toggle, but not
the name filter.
Closes #377.

Signed-off-by: Yannick Schaus <github@schaus.net>