[WIP/RfC] Developer Toolbar
Allow administrators to open a sidebar with tools they might like to
have handy when they're working for instance on a rule or a page.

The developer sidebar is only available from the developer tools menu
or anywhere with the Shift+Alt+D keyboard shortcut, if all these
conditions are met:

1. the user has an administrator role;
2. the device is a desktop (or laptop, that is, not a phone or tablet);
3. the width of the screen is >= 1280px.

Only the first tool "Item Monitor" is barely functional for now.

Signed-off-by: Yannick Schaus <github@schaus.net>