Dark mode, horizontal layout, switch between code & Blockly
Toolbox in horizontal layout on mobile devices by default
Disable trashcan
Allow to switch a application/javascript script to Blockly
if it is empty - or back to raw code editing if the
Javascript generated by Blockly is empty.

Switch between code preview (read-only) & Blockly
with a FAB button, or Ctrl-B on the keyboard.

Catch code generation exceptions.

Signed-off-by: Yannick Schaus <github@schaus.net>