Home page model cards rewrite
Refactored the experimental-grade code with some mixins,
better separated code.

Implements #555:
Allow the home page cards to be customized - reordered, grouped,
and edited, with config & slots and a dedicated editor.

Implements #556:
Add glance badges to the location cards (other types tbd.)
See #556 for the specifications on how to organize your model
to get the badges.

Styling fixes and cleanups.

Signed-off-by: Yannick Schaus <github@schaus.net>