Widgets: add oh-repeater, dayjs to expression context
This adds a new oh-repeater to the system library, allowing
to iterate over arbitrary arrays, ranges, or items in a certain
group or having certain tags, with support for filtering and
mapping. The children components in the default slot will be
repeated for each iteration and the current element, along with
the index and source array, will be provided as variables to the
descendants.

Add dayjs to the expression context for powerful date manipulation
and formatting. The default locale will be dynamically loaded
from the current openHAB locale.
The following Day.js plugins are loaded:
- relativeTime
- calendar
- localizedFormat
- isoWeek
- isToday
- isYesterday
- isTomorrow
Other can be added on demand.

Try to find a f7-component relative to the oh-component and merge
their parameters during hinting.

Signed-off-by: Yannick Schaus <github@schaus.net>