[WIP] i18n infrastructure + end-user facing pages
Related to #394.

Add vue-i18n.
The localized messages are in `@/assets/i18n/{domain}/{locale}.json`,
and are localized by component.
They should be compatible with Crowdin.
They are also referenced in single-file components with this trick:
```
import { loadLocaleMessages } from '@/js/i18n'

export default {
  ...
  i18n: {
    messages: loadLocaleMessages(require.context('@/assets/i18n/{domain}'))
  }
```
That way they should be code splitted away from the entry point bundle
when their related component is, and they are scoped to that component.

The `common` domain with a few common and home page related messages
are loaded on initialization.

English and French messages are provided, as well as German (to be
proofed!) for the `common` domain.

A footer message is added to the settings & developer tools menu if
the locale has common messages to inform the user that these parts are
not translated yet.

Signed-off-by: Yannick Schaus <github@schaus.net>