Add wizards to help adding rule modules
Replace the default flat grouped list of module types by
more a user-friendly UI that is aware of the core modules
and offers an improved experience.

Examples include directly jump into the script editor, in
Blockly mode, or pick the item from the model before
deciding which event to consider.

Also now the behavior of the icons on the cron triggers
and script actions/conditions is now reversed: clicking on
the bar will launch the special action, and clicking on the
little icon will bring up the generic module editor popup.

Fixed time of day parameter, replaced day of week by an
inline list.

Moved the Blockly button to the bottom in the script editor.

Fixed mistakes in some module titles suggestions.

Closes #610.

Signed-off-by: Yannick Schaus <github@schaus.net>