basic ui: Improve consistency of escaping state and category values
As opposed to papering over inconsistent encoding behavior with
decoding in the frontend, here's an attempt to try to clean up our
encoding behaviors just a bit.

For %state%, this was split into two template variables, %state% that
is only HTML-escaped, and %stateUrlEnc% for use in the icon URLs. A
similar convention was adopted for %category%, which is also used in
URLs.

The goal is that, by convention, data- attribute values should _never_
be URL-encoded, and rather just safely escaped for HTML. A separate
url-encoded template variable can be offered if needed. Hopefully this
will make the markup easier to reason about in the frontend context,
and avoid future encoding issues.

Signed-off-by: Lucas Christian <lucas@lucasec.com>