Avoid the colorpicker entering endless loops
As reported in https://community.openhab.org/t/wiki-building-pages-in-the-oh3-ui-documentation-draft-1-3/104382/58
Sometimes the colorpicker enter infinite loops
sending the same command endlessly, due to
either rounding errors in the HSB components
or other causes.

These fixes should make it better and avoid
these situations.

Add a key to the widget preview in the model
details to avoid reusing widgets when the
configuration differs.

Signed-off-by: Yannick Schaus <github@schaus.net>