Disable slider in item action module wizard
The rule designer's new action module wizard
has a slider to help set the command for dimmers,
rollershutters & numbers but it doesn't work well
and prevents the input of command values
outside the range, which might prove problematic
(in particular there's no way anymore to set a
rollershutter item to UP/DOWN/STOP...).
So disable it for now.

Also fix bugs with the "thing" trigger wizard.

Signed-off-by: Yannick Schaus <github@schaus.net>