Auth system refactoring
This brings a new number of improvements regarding auth:

- The SSE connections will be made using _event-source-polyfill_
which allows setting headers on the request, and setting
authorization headers or tokens as required has been added.
The main UI can now be used when the API is fully protected,
including the user operations like controlling items ("Implicit
user role for unauthenticated requests" disabled in Settings >
API Security). This closes #699.

- The UI now detects when the first /rest/ API calls fails with
a 401 Unauthorized error. This can happen when it's installed
as a standalone PWA ("Add to Home Screen" or "Install" in Chrome).
In that case, it will display a f7 login dialog and will also try
to use the Credentials API to store the passed credentials in
the browser, and retrieve them back to log in automatically the
next time. It will use these as Basic authentication credentials.
This prevents "empty screen" errors when using e.g.
myopenhab.org and the cloud credentials couldn't be passed.
The Basic credentials for proxy services can also be passed from
mobile with the following Javascript functions in the `OHApp`
object:
  * `string getBasicCredentialsUsername()`
  * `string getBasicCredentialsPassword()`

- Detects whether the API auth bundle is disabled (no /auth
endpoint), hides the authorize button and unlocks the admin menus.

Signed-off-by: Yannick Schaus <github@schaus.net>