Fix analyzer back button if another popup was opened
Framework7 would not consider an async route to be a modal route,
leading to critical code not being executed and thus causing #1068.
This regression was introduced when the routes were changed to async in #948.

So introduce a new component and lazy load the analyzer inside it, so
there's a minimal impact on the app entry point.

Changed some z-indexes to make sure the popup backdrop is displayed between
the analyzer popup and the item picker popup.

Fixes #1068.
Fixes #1067.

Signed-off-by: Yannick Schaus <github@schaus.net>