Replace Add-on Management UI with Add-on Store (#1158)
Depends on https://github.com/openhab/openhab-core/pull/2405.

This replaces the current section of the UI for managing
add-ons with a completely new "store-like" interface.
External data, like logos and detailed descriptions for
distribution add-ons are fetched from well-known locations
on the website or GitHub to provide a "store" look & feel.

The add-on store is orgnanized around 4 categories: bindings,
automation, UI, and others, with a separate search tab.
Each category is splitted into subsections. The availability
of these sections depend on the add-on services that are
registered. The sections currently supports add-ons provided
by the "karaf" & "marketplace" services, as provided by
https://github.com/openhab/openhab-core/pull/2405.
Other add-on services should be supported as well.

Signed-off-by: Yannick Schaus <github@schaus.net>