Fix Roboto fontface path
At least in the latest version of roboto-fontface module, roboto medium and regular fonts are placed in ./node_modules/roboto-fontface/fonts/roboto/ directory not in ./node_modules/roboto-fontface/fonts/ so the paths are not found and Roboto-Medium.* and Roboto-Regular.* fonts are not copied into ./src/main/resources/web directory.
This is causing issue #1273.

Signed-off-by: Silviu Chingaru <silviuchingaru@yahoo.com>