Add possibility to edit JSON variable keys in widgets (#1556)
Implements a possibility to work with nested JSON objects/arrays in widget variables. It allows to set the value of a custom key of the variable. The key can be nested in any depth to allow more complex scenarios. 

Adds a new action configuration `actionVariableKey` which allows to set the key of the object. The configuration syntax is like this:
```
actionVariableKey: name
actionVariableKey: user.name
actionVariableKey: user[0].address[1].street
actionVariableKey: [0]
actionVariableKey: [0].label
```

If you want to edit the key `user.name` but the key `user` is missing in the object, the parent hierarchy is automatically created. (Same applies if you want to add an value to an array but the array is not existing) This should make it easier for not so advanced users.

Signed-off-by: Florian Michel <florianmichel@hotmail.de>
Also-by: Yannick Schaus <github@schaus.net>