Add step parameter to oh-input (#1726)
When using number type in an oh-input card or oh-input-item list item, a
default step value of 1 is always respected. This leads to confusion
with users not being able to enter fractional numbers.
See [forum
post](https://community.openhab.org/t/problems-with-numerical-values-in-input-cards/134774).

This PR makes the step parameter visible, and uses a default of any
value (no step check) if no step value is provided. This is a small
breaking change, but less confusing to users.

--
Signed-off-by: Mark Herwege <mark.herwege@telenet.be>