[BasicUI] Remove support for List element
Related to openhab/openhab-android#3285

This sitemap widget is not documented.

It is not implemented neither in Android app nor iOS app.

It was partially implemented in BasicUI but does not work well (it breaks the layout).

Without any opened issue, we can be relatively sure that noone is using it.

Signed-off-by: Laurent Garnier <lg.hc@free.fr>