Add recommended add-ons to the setup wizard
Closes #1659.

This adds the ability to define a list of recommended add-ons which will be pre-selected in the setup wizard.

Currently, JS Scripting (GraalJS) Automation as a Blockly Dependency and RRD4J Persistence as default persistence service (used e.g. for trend-line, analyzer in the UI) are selected as recommended add-ons.

A note about recommended add-ons in general is displayed.

Signed-off-by: Florian Hotze <florianh_dev@icloud.com>