`oh-slider`: Fix sending command when state update with decimals is received
Fixes #1862.

When a state update like 40.545 was received, the slider component commanded the Item to 41 because of the rounding used.
Unfortunately, the slider component has no built-in way to avoid that.
This decouples the slider value from the actual Item value and only send command when the difference between them is larger or equal than step size.

Signed-off-by: Florian Hotze <florianh_dev@icloud.com>