Add a persistence configuration page
It is accessible from the add-on settings page and has both a design and a code tab.

The design tab allows to set persistence strategies for Items, define cron strategies and set the default strategies.
It does not duplicate names for (cron) persistence strategies and duplicate configs for the same set of Items.

The code tab also allows to also specify threshold and time filters and needs minor adjustment once https://github.com/openhab/openhab-core/pull/3642 is merged code completion is not provided.

Signed-off-by: Florian Hotze <florianh_dev@icloud.com>