Items list: Fix searchbar broken after Items list update
Fixes #1334.
Fixes #1702.

After editing or removing an Item, the searchbar of the Items list was not working anymore.
Console showed this error: `TypeError: undefined is not an object (evaluating 'n.virtualList.params')`

It seems that this was caused because the replaceAllItems operation was applied on the virtual list, but the was searchbar not re-rendered afterward.

Signed-off-by: Florian Hotze <florianh_dev@icloud.com>