Rule/Script edit: Fix not editable not properly handled & Always show tags (#2276)
Fixes an issue where the configuration of a script action of a
file-based rule (e.g. those created using JSRule or Rule Builder of
openhab-js) was shown as editable, even though the rule is not editable
because it is provisioned from file.

Always show the tag input, even if there are no rule tags. This is more
consistent and avoids confusion, as you can now clearly see that a rule
has no tags.

Signed-off-by: Florian Hotze <florianh_dev@icloud.com>