Widget expressions: Partially avoid requesting Item state if Item name is undefined
When using the `@` and `@@` shortcuts for Item displayState and state in widget expressions,
properly Item name undefined.
Unfortunately I did not find a way to improve that for `items[itemName]` expressions.

Requesting the Item state from the server when the Item name is undefined leads to warnings in the log:
```
[WARN ] [se.internal.SseItemStatesEventBuilder] - Attempting to send a state update of an item which doesn't exist: undefined
```

Signed-off-by: Florian Hotze <florianh_dev@icloud.com>