[BasicUI] Move most of app settings to per-device (browser) settings
Closes #1832
Closes #1930
Closes #2342

A new page is available to show and update all per-device settings. This page is accessible through the sitemap list page or the home page of any sitemap.
Basic UI theme now matches the dark mode of Main UI. It can be updated either in Main UI or in the new page of Baisc UI.
Web Audio can also be enabled either in Main UI or Basic UI.

The iconify setting is now enabled by default. The script to handle iconify icons is now loaded in sitemap pages only if the iconify setting is enabled.

New settings are also available to control the size of the chart and the size of content (text, line) in the chart.

A new setting is also available to enable a bigger font size.

Signed-off-by: Laurent Garnier <lg.hc@free.fr>