Thing edit: Fix changing ID shows wrong value & Always show UID (#2475)
Fixes #2469.

After changing the ID of a Thing (only possible when creating a new
Thing), the displayed value was reset to the pre-filled default when
moving to a different field. When saving the Thing, the entered value
was used. This was because Vue did not properly notice the change of the
ID.

Signed-off-by: Florian Hotze <florianh_dev@icloud.com>