[BasicUI] Add special handling for "none" icon
Same behaviour as in the Android app is now implemented for the "none" icon.
This icon means "no icon" and we don't want to request the icon servlet.
When this icon is associated to a sitemap element, the icon servletis no more requested.
Instead, a simple empty SVG tag is considered and inserted in the HTML page.
There is also no dynamic behaviour (with item state) applied for this icon.

Additionally, in case the loading of any OH icon fails, the IMG tag is replaced by an empty SVG tag in the HTML page.

Fix openhab/openhab-core#3977

Signed-off-by: Laurent Garnier <lg.hc@free.fr>