[MainUI] SIP Auto answer/dial features
This adds two new features:

The first is to auto accept calls if they match a configured list of numbers (* for all, 'user', or 'user@host` in a comma delimited list)

The second is to auto dial a pre-configured number when the client finishes registering (or when the client connects if registration is disabled).

Both can be safely used together.  This allows for features like intercoms, as well as widgets that connect when popped up.

Signed-off-by: Dan Cunningham <dan@digitaldan.com>