Home page: Add option to make simple navbar the default on mobile
This adds a two new boolean config parameter to the home page settings,
which allow to set the simple navbar as default on mobile and desktop devices.
The navbar style setting in the about page has been changed from a toggle to a segmented button with the options default, simple, large.

Signed-off-by: Florian Hotze <florianh_dev@icloud.com>